+++ packages/ui_v2/src/components/commonV2/Stepper/UtilityButton.tsx
@@ -2,12 +2,19 @@ import React from "react";
 
 interface Props {
   title: string;
-  callback: (i: boolean) => void;
+  openModal: (i: boolean | string) => void;
 }
 export function UtilityButton(props: Props) {
-  const { title, callback } = props;
+  const { title, openModal } = props;
+  const handledCallback = () => {
+    if (title === "Import CSV") {
+      openModal(title);
+    } else {
+      openModal(true);
+    }
+  };
   return (
-    <button style={styles.button} onClick={() => callback(true)}>
+    <button style={styles.button} onClick={handledCallback}>
       {title}
     </button>
   );
